Refactor unit tests for potential and pair modules
