+++ tests/simulate/test_dilute.py
@@ -6,7 +6,7 @@ import numpy as np
 
 import relentless
 
-from ..potential.test_potential import LinPot
+from ..potential.test_pair import LinPot
 
 class test_Dilute(unittest.TestCase):
     """Unit tests for relentless.simulate.Dilute"""
