+++ tests/simulate/test_hoomd.py
@@ -7,7 +7,7 @@ import hoomd
 import numpy as np
 
 import relentless
-from ..potential.test_potential import LinPot
+from ..potential.test_pair import LinPot
 
 class test_HOOMD(unittest.TestCase):
     """Unit tests for relentless.HOOMD"""
